Rename traceback -> backtrace to avoid naming conflict with global module
