+++ src/scout_apm/core/backtrace.py
